Mapping
The mapping transformation creates a collection from the results of a function on the elements of another collection.
The basic mapping function is map().
It applies the given lambda function to each subsequent element and returns the list of the lambda results.
The order of results is the same as the original order of elements.